Fix compiler warning.
git-svn-id: file:///.data/svn_convert/svn@6351 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/plugins/ontop/ontop.cpp
@@ -101,7 +101,7 @@ OnTopPlugin::OnTopPlugin(unsigned base, Buffer *config)
 
 OnTopPlugin::~OnTopPlugin()
 {
-    Event eCmd(EventCommandRemove, (void*)CmdOnTop);
+    Event eCmd(EventCommandRemove, (void*)(unsigned long)CmdOnTop);
     eCmd.process();
 
     setOnTop(false);
